Update "is" and "addClass" to handle invalid params
In spite of testing for undefined selection in the "is" function it was still possible to pass an invalid selector that was causing an unhandled promise exception. It turned out to be easiest to handle with a simple try catch here as I couldn't see a way to do so without pattern matching for invalid selectors which seemed like a lot overhead. Once this was finally resolved there was an issue with addClass function not checking for empty strings. It seems like some discussion on perhaps a global error catching strategy for this file might be worth discussing. In my case this was a fairly serious issue as the Googlebot mobile indexing just completely choked and only partially loaded the site really hurting mobile indexing.